% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCatagsFile.R
\name{createCatagsFile}
\alias{createCatagsFile}
\title{Create categories file}
\usage{
createCatagsFile(path = NULL)
}
\arguments{
\item{path}{File path of where you wish the categories file to be saved}
}
\value{

}
\description{
Set up a category file that then can be used with the creatTC function.
}
\examples{
# link <- getwd()
# createCatagsFile(path = link)
}
