% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subFunctions.R
\name{fixTimes}
\alias{fixTimes}
\title{Fix excel timing issues}
\usage{
fixTimes(date, endStart)
}
\arguments{
\item{date}{Date parameter}

\item{endStart}{Start or End time}
}
\value{
A POSIXct date-time is exported
}
\description{
This function fixes some of the excel related timing problems, date and star/end times are merged and converted to class POSIXct
}
\examples{
fixTimes(Sys.Date(), '11:00')
}
