% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subFunctions.R
\name{getRecentMondays}
\alias{getRecentMondays}
\title{Get recent Mondays}
\usage{
getRecentMondays()
}
\description{
This function looks back over the past 70 days and draws out dates that fall on Monday.
}
\examples{
getRecentMondays()
}
