% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subFunctions.R
\name{is.integer0}
\alias{is.integer0}
\title{Find whether data are integer(0)}
\usage{
is.integer0(x)
}
\arguments{
\item{x}{Data of interest}
}
\description{
Find whether data are integer(0)
}
\examples{
data <- 1
is.integer0(data)
}
