% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeOTLs.R
\name{mergeOTL}
\alias{mergeOTL}
\title{Compile OTL data}
\usage{
mergeOTL(folderOTL, category, quarterYear = "all", aggregate = FALSE)
}
\arguments{
\item{folderOTL}{The folder you have saved the OTL export files}

\item{category}{Filepath to the categories file}

\item{quarterYear}{Set to 'all', this compiles all OTLs. If you set to 'current' it will compile all the OTLs that are in the same quarter as sys.Date(). To specify a quarter set to the format of 'Qx yyyy' e.g. 'Q1 2023'.}

\item{aggregate}{Set to FALSE. If TRUE it will sum all the ours spent on each category.}
}
\value{
All OTL files merged into one table
}
\description{
Compile OTL data
}
\examples{

folder <- 'C:/Users/jpayne05/OneDrive - Defra/Time_Recording/OTLs'
cats <- 'C:/Users/jpayne05/OneDrive - Defra/Time_Recording/Categories/Categories_TCs.csv'
dt <- mergeOTL(folderOTL = folder, category = cats, quarterYear = 'all')
dt
dt1 <- mergeOTL(folderOTL = folder, category = cats, quarterYear = 'all',
                aggregate = TRUE)
dt1
}
