% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.totalOTLs}
\alias{plot.mergedOTLs}
\title{Plotting OTL data}
\usage{
plot(x, ...)

\method{plot}{totalOTLs}(x, ...)

\method{plot}{mergedOTLs}(x)
}
\arguments{
\item{x}{Dataset derived from mergeOTL() function}

\item{...}{Other options passed to ggplot(., aes())}
}
\value{
ggplot bar chart to detail where OTL is falling
}
\description{
Plotting OTL data
}
\examples{
\dontrun{
plot(dt1)
}
}
