% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subFunctions.R
\name{toHMS}
\alias{toHMS}
\title{Convert times to hms (hours minutes seconds)}
\usage{
toHMS(x, na.rm = FALSE)
}
\arguments{
\item{x}{Data of interest}

\item{na.rm}{Set as TRUE}
}
\value{
HMS times
}
\description{
Convert times to hms (hours minutes seconds)
}
\examples{
# Leave blank
}
