% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subFunctions.R
\name{toHours}
\alias{toHours}
\title{Convert hms to seconds}
\usage{
toHours(x, na.rm = FALSE)
}
\arguments{
\item{x}{Data of interest}

\item{na.rm}{Set as TRUE}
}
\value{
Seconds in a numeric form
}
\description{
Convert hms to seconds
}
\examples{
# Leave blank
}
