% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subFunctions.R
\name{toTime}
\alias{toTime}
\title{Convert excel "time" into R compatible time}
\usage{
toTime(x, na.rm = FALSE)
}
\arguments{
\item{x}{Data of interest}

\item{na.rm}{Set as TRUE}
}
\value{
Converted excel time to R time
}
\description{
Convert excel "time" into R compatible time
}
\examples{
# Leave blank
}
