% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTC.R
\name{createTC}
\alias{createTC}
\title{Create Time card}
\usage{
createTC(
  categories = NULL,
  daily = NULL,
  outCal = NULL,
  week_start = NULL,
  split = NULL,
  weight = NULL,
  pathOTL = NULL,
  export = TRUE
)
}
\arguments{
\item{categories}{Link to your categories file (csv)}

\item{daily}{Link to your daily hours (xlsx)}

\item{outCal}{Link to your Outlook export file (csv)}

\item{week_start}{The week of interest}

\item{split}{A selection of projects to split time between}

\item{weight}{Weighting factor for each project}

\item{pathOTL}{Link to folder where you wish the OTL data to be exported to}

\item{export}{Set as True - saves the OTL form in your pathOTL folder}
}
\value{
A time card in csv format
}
\description{
Function that links daily hours to outlook calendars to produce an OTL form
}
\examples{

\dontrun{
catags <- 'C:/Users/jpayne05/Desktop/Time/Categories_TCs.csv'
dailHours <- 'C:/Users/jpayne05/Desktop/Time/Daily_hours.xlsx'
outlC <- 'C:/Users/jpayne05/Desktop/Time/calendar_appoints3.csv'
weekS <- '2023-05-22'
pathOTL <- 'C:/Users/jpayne05/Desktop/Time'
tasks <- c('Cap Skills', 'FFIDP', 'Reactive Forecasting')
weightings <- c(1, 2, 1)
tc <- createTC(categories = categories,
               daily = daily,
               outCal = outCal,
               week_start = week,
               split = tasks,
               weight = weightings,
               export = FALSE)
tc
}
}
