% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOTLtotals.R
\name{plot.totalOTLs}
\alias{plot.totalOTLs}
\title{Plotting OTL data}
\usage{
\method{plot}{totalsOTL}(x, ...)
}
\arguments{
\item{x}{Dataset derived from mergeOTL() function}

\item{...}{Other options passed to ggplot(., aes())}
}
\value{
ggplot bar chart to detail where OTL is falling
}
\description{
Plotting OTL data
}
\examples{
\dontrun{
plot(dt1)
}
}
